feat: Add SparseImageStream and export helpers
Introduce SparseImageStream (LibSpraseSharp/SparseImageStream.cs) to map sparse-file blocks into a full sparse-format stream and add SparseFile methods GetExportStream and GetResparsedStreams to export or split sparse images into multiple streams (honors max file size and optional CRC). Also apply small cleanups: remove unnecessary casts in LibLpSharp/Utility.cs and LpDump/Program.cs (bit-shift/offset calculations) and simplify theme initialization in SettingsViewModel (PartitionToolSharp.Desktop/ViewModels/SettingsViewModel.cs).

+++ LibLpSharp/Utility.cs
@@ -65,7 +65,7 @@ public static class Utility
                 return new FilesystemInfo
                 {
                     Type = "EROFS",
-                    Size = (ulong)blocks << (int)log2_blksz
+                    Size = (ulong)blocks << log2_blksz
                 };
             }
